+++ R/interPlot.R
@@ -1,69 +1,61 @@
 #' interPlot
 #'
 #' @description Interactive 3D plot by calling `plotly` package functions.
-#' @usage interPlot(data,xvar,yvar,zvar,wvar=NULL,numGrpx=10,numGrpy=10)
-#' @param data a data frame.
-#' @param xvar either an integer to specify the position of the variable in the data frame, or the name of the variable.
-#' @param yvar either an integer to specify the position of the variable in the data frame, or the name of the variable.
-#' @param zvar either an integer to specify the position of the variable in the data frame, or the name of the variable.
-#' @param wvar either an integer to specify the position of the variable in the data frame, or the name of the variable.
+#' @usage interPlot(x,y,z,weight=NULL,exposure=NULL,numGrpx=10,numGrpy=10,
+#' xname="x",yname="y",zname="z")
+#' @param x,y a vector indicates the dependent variable that you want to visulise on (i.e. Age)
+#' @param z a vector indicates the response variable that you want to visulise on (i.e. Age)
+#' @param weight Optional. A numerical vector to specify the weights used for calculating weighted average of response.
+#' normally this is the figures from over/down-samping.
+#' @param exposure Optional. A numerical vector to specify the exposure used for calculating weighted average of response.
 #' @param numGrpx,numGrpy integer. Give the number of buckets for either x or y factors.
+#' @param xname,yname,zname Optional. Characters to be shown on plot.
 #' @details This functions gives a 3D interactive view of between factors.
 #' This is really useful when modeller wants to assess the interaction terms.
-#' Hence the "zvar" can be actual response data or model predictions.
+#' Hence the `z` can be actual response data or model predictions.
 #' @author Sixiang Hu
 #' @importFrom data.table data.table setkey key := dcast
-#' @importFrom plotly plot_ly layout %>%
+#' @importFrom plotly plot_ly layout add_surface %>%
 #' @export interPlot
 #' @examples
 #' 
-#' interPlot(mtcars,"wt","mpg","vs")
+#' interPlot(mtcars$disp,mtcars$mpg,mtcars$qsec)
 
-interPlot <- function(data,xvar,yvar,zvar,wvar=NULL,numGrpx=10,numGrpy=10){
+interPlot <- function(x,y,z,weight=NULL,exposure=NULL,numGrpx=10,numGrpy=10,
+                      xname="x",yname="y",zname="z"){
   # Error Trap
-  if( .isDFnull(data) ) stop("data set provided is null.")
-  if( is.null(xvar) ) stop("X variable provided is null.") 
-  if( is.null(yvar) ) stop("Y variable provided is null.")
-  if( is.null(zvar) ) stop("Z variable provided is null.")
+  if (is.null(x)) stop("x provided is blank.")
+  if (is.null(y)) stop("y provided is blank.")
+  if (is.null(z)) stop("z provided is blank.")
   
-  # Find data column
-  posi <- .VarPosition(data,xvar)
-  x <- data[[posi$posi]]
-  if ( (is.numeric(x) || is.integer(x) ) && nlevels(as.factor(x))>=100 ) {
+  nr <- length(x)
+  
+  if ( is.numeric(x) && nlevels(as.factor(x))>=100 ) {
     new_band <- dmBreak(x,numGrpx)
     x <- cut(x,new_band,include.lowest = TRUE,ordered_result = TRUE)
   }
 
-  posi <- .VarPosition(data,yvar)
-  y <- data[[posi$posi]]
-  if ( (is.numeric(y) || is.integer(y) ) && nlevels(as.factor(y))>=100 ) {
+  if ( is.numeric(y) && nlevels(as.factor(y))>=100 ) {
     new_band <- dmBreak(y,numGrpy)
     y <- cut(y,new_band,include.lowest = TRUE,ordered_result = TRUE)
   }
 
-  posi <- .VarPosition(data,zvar)
-  z <- data[[posi$posi]]
+  if(is.null(weight)) weight <- rep(1,nr)
+  if(is.null(exposure)) exposure <- rep(1,nr)
   
-  if( is.null(wvar)) {
-    w <- rep(1,dim(data)[1])
-  }
-  else {
-    posi <- .VarPosition(data,wvar)
-    w <- data[[posi$posi]]
-  }
-  
-  dt <- data.table::data.table(x = x,y = y,z = z,w = w)
+  dt <- data.table::data.table(x = x,y = y,z = z,w = weight,e=exposure)
   data.table::setkey(dt,x,y)
-  
-  dt2  <- dt[,lapply(.SD,weighted.mean,w=w),by=list(x,y),.SDcols=c("z","w")]
-  dt2 <- data.table::dcast(dt2,x~y,value.var="z",drop=FALSE)
-  
-  zvalue <- suppressWarnings(DataMan::PopMiss(as.matrix(dt2[,-1,with=FALSE]),
-                                     na.treatment = "replace",0))
 
-  plotly::plot_ly(x=~x,y=~y,z=~zvalue,type = "surface") %>%
-    plotly::layout(scene=list(xaxis=list(title=paste0(xvar," (x)")),
-                           yaxis=list(title=paste0(yvar," (y)")),
-                           zaxis=list(title=paste0(zvar," (z)")))
+  dt2 <- dt[,lapply(.SD,function(x,w,e) sum(x*w,na.rm = TRUE)/sum(e*w,na.rm = TRUE),e=e,w=w),by=list(x,y),.SDcols="z"]
+  dt2 <- as.matrix(data.table::dcast(dt2,x~y,value.var="z",drop=FALSE,fill=0)[,-1])
+
+  plotly::plot_ly(z=~dt2) %>%
+    plotly::add_surface() %>%
+    plotly::layout(scene=list(xaxis=list(title=paste0(xname," (x)")),
+                           yaxis=list(title=paste0(yname," (y)")),
+                           zaxis=list(title=paste0(zname," (z)")))
                    )
 }
+
+#global variable
+globalVariables(c("e","w"))